+++ package.json
@@ -4,7 +4,7 @@
   "description": "A Vue.js plugin for getting CSRF-token",
   "main": "dist/index.js",
   "scripts": {
-    "build": "npx babel src/index.js -o dist/index.js"
+    "build": "npx babel src -d dist"
   },
   "keywords": [
     "vue",
@@ -13,7 +13,8 @@
   "author": "Mikita Amialkovich",
   "license": "MIT",
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-preset-es2015": "^6.24.1"
+    "@babel/cli": "^7.4.4",
+    "@babel/core": "^7.4.5",
+    "@babel/preset-env": "^7.4.5"
   }
 }
