+++ src/Csrf.js
@@ -1,0 +1,13 @@
+module.exports = class Csrf {
+  constructor(token) {
+    this._token = token;
+  }
+
+  get() {
+    return this._token;
+  }
+
+  set(token) {
+    this._token = token;
+  }
+}
